# src/main/resources/api/api.yml
openapi: 3.0.3
info:
  version: '1.0'
  title: Company Management API
  description: |
    API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux,
    mouvements financiers, dépenses, achats et frais de transport.
    Basée sur le schéma de base de données fourni.
servers:
  - url: 'https://api-dev.company.com'
  - url: 'https://api-prod.company.com'
security:
  - BearerAuth: [ ]
paths:
...
  '/companies/{comp_id}/job/{job_id}/user/{user_id}/expenses/{expenses_id}/travel_expenses/{travel_expenses_id}/travel_materials/{id}':
    get:
      tags:
        - TravelMaterials
      summary: Get travel materials by identifier
      operationId: getTravelMaterialsById
      parameters:
        - name: comp_id
          in: path
          required: true
          schema:
            type: string
        - name: job_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: expenses_id
          in: path
          required: true
          schema:
            type: string
        - name: travel_expenses_id
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The identified travel materials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelMaterials'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - TravelMaterials
      summary: Delete travel materials by identifier
      operationId: deleteTravelMaterialsById
      parameters:
        - name: comp_id
          in: path
          required: true
          schema:
            type: string
        - name: job_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: expenses_id
          in: path
          required: true
          schema:
            type: string
        - name: travel_expenses_id
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Travel materials deleted successfully
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /companies/{comp_id}/job/{job_id}/user/{user_id}/expenses/{expenses_id}/travel_expenses/{travel_expenses_id}/travel_materials:
    get:
      tags:
        - TravelMaterials
      summary: Get all travel materials
      operationId: getTravelMaterials
      parameters:
        - name: comp_id
          in: path
          required: true
          schema:
            type: string
        - name: job_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: expenses_id
          in: path
          required: true
          schema:
            type: string
        - name: travel_expenses_id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
        - name: travel_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of travel materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TravelMaterials'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - TravelMaterials
      summary: Create new travel materials or update existing ones
      operationId: crupdateTravelMaterials
      parameters:
        - name: comp_id
          in: path
          required: true
          schema:
            type: string
        - name: job_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: expenses_id
          in: path
          required: true
          schema:
            type: string
        - name: travel_expenses_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CrupdateTravelMaterials'
      responses:
        '200':
          description: The created or updated travel materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TravelMaterials'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'


  # =========================
  # TRAVEL EQUIPMENT
  # =========================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '429':
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    AuditFields:
      properties:
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
        updated_by:
          type: string
        comment:
          type: string

    # =========================
    # USER
    # =========================
...
    TravelMaterials:
      allOf:
        - $ref: '#/components/schemas/AuditFields'
      properties:
        id:
          type: string
        travel_id:
          type: string
        material:
          type: string
        quantity:
          type: integer
        quantity_received:
          type: integer
    CrupdateTravelMaterials:
      properties:
        id:
          type: string
        travel_id:
          type: string
        material:
          type: string
        quantity:
          type: integer
        quantity_received:
          type: integer

    # =========================
    # TRAVEL EQUIPMENT
    # =========================
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error